--- admin/info.php
@@ -1,7 +1,0 @@
-<?php
-
-require_once 'bootstrap.php';
-
-$info = new Info($config);
-$info->execute();
-echo $info->_log_messages;
