--- admin/pull.php
@@ -1,7 +1,0 @@
-<?php
-
-require_once 'bootstrap.php';
-
-$pull = new Pull($config);
-$pull->execute();
-echo $pull->_log_messages;
