+++ arse_admin/info.php
@@ -1,0 +1,7 @@
+<?php
+
+require_once 'bootstrap.php';
+
+$info = new Info($config);
+$info->execute();
+echo $info->_log_messages;
